Fix warning

+++ apps/ro-localization/CPosePDFParticlesExtended.cpp
@@ -1236,7 +1236,7 @@ double  CPosePDFParticlesExtended::auxiliarComputeObservationLikelihood(
 			obserDumm = *obs;
 
 			// Introduce bias:
-			ASSERT_( obserDumm.sensedData.size() == x->state.size() );
+			ASSERT_( (int)obserDumm.sensedData.size() == (int)x->state.size() );
 			for (size_t k=0;k<size_t(obserDumm.sensedData.size());k++)
 				obserDumm.sensedData[k].sensedDistance -= x->state[k];
